Added an option for second-order vertical remapping
The transport driver of the Glissade dycore has three parts:
(1) horizontal transport (IR or upwind), (2) application of surface mass balance,
and (3) vertical remapping back to sigma coordinates.
Until now, step (3) has been first-order accurate in the z coordinate.
When part of one layer is merged into an adjacent layer, the tracer
value in that layer has been taken to be the mean value.

With this commit I've modified the vertical remapping subroutine in
glissade_transport.F90 to allow second-order-accurate remapping.
Within each layer, the tracer value is reconstructed as a linear field,
with the gradient limited to preserve monotonicity.  As a result,
layers can be remapped in the vertical with less numerical diffusion.

I added an new config option, which_ho_vertical_remap. The values are
  integer, parameter :: HO_VERTICAL_REMAP_FIRST_ORDER = 0
  integer, parameter :: HO_VERTICAL_REMAP_SECOND_ORDER = 1
It would be more intuitive, perhaps, to use '1' for first-order and '2'
for second-order, but this would break the CISM convention that there
is always a '0' option.

This option is part of the 'options' derived type and is now passed into
glissade_transport_driver.  First- and second-order remapping are handled
by the same new subroutine.  For now I left the old subroutine in place for
reference, but I will remove it in a later commit.

The default for now is first-order.  I verified for the dome problem that results
are BFB with first-order remapping.

For second-order remapping, the changes are modest but not negligible.
In a standard EISMINT-2 test run for 20 ka, I found that temperature at a
given layer can vary by ~0.5 degree, and ice age by ~1000 years, between
first-order and second-order runs.