Added a new 'ocean-cavity' option for the flotation function
For which_ho_ground = 0 (no GLP) or 1 (GLP), the Glissade dycore computes a
flotation function at cell centers and interpolates it to cell vertices.
The value at or near the vertex determines the grounded ice fraction at the
vertex, which in turn multiplies the basal traction (beta). Thus, whether a vertex
is deemed to be grounded or floating can make a big difference to the dynamics.

Originally, there was just one option, f_flotation = -(rhow*b)/(rhoi*H), aka f_pattyn.
With the commit before this one, I added an option f_flotation = (rhoi*H)/(-rhow*b), aka inverse pattyn.
With this commit I added a third option, f_flotation = -rhow*b - rhoi*H.
This is called the ocean-cavity option because f_flotation, if positive, is equal
to the depth of the ocean cavity beneath an ice shelf.  Thanks to Xylar Asay-Davis
for suggesting this function.  He has found it to be better-behaved than either of the
other two functions when working with Bedmap2 data.  It is linear in both b and H,
unlike f_pattyn and its inverse.  Its value is 0 at the GL, negative for grounded ice,
and positive for floating ice.

It remains to be seen whether the new ocean-cavity option gives consistently stable flow in
situations where f_pattyn allows large and potentially CFL-violating velocities. I tested it on a
Greenland 4-km grid that is known to be prone to large velocities, and it ran stably
for 1 year with a time step of 0.05 yr.  Results are very similar to those for 1/f_pattyn
in most of the ice sheet. But I have kept 1/f_pattyn as the default pending further testing.

To switch between the 3 options, the user can now set which_ho_flotation_function
in the config file without re-compiling the code.  The options are as follows:

  integer, parameter :: HO_FLOTATION_FUNCTION_PATTYN = 0
  integer, parameter :: HO_FLOTATION_FUNCTION_INVERSE_PATTYN = 1
  integer, parameter :: HO_FLOTATION_FUNCTION_OCEAN_CAVITY = 2

  (0) f_flotation = (-rhow*b/rhoi*H) = f_pattyn; <=1 for grounded, > 1 for floating
  (1) f_flotation = (rhoi*H)/(-rhow*b) = 1/f_pattyn; >=1 for grounded, < 1 for floating
  (2) f_flotation = -rhow*b - rhoi*H = ocean cavity thickness; <=0 for grounded, > 0 for floating