Imposed a maximum surface slope for Glissade; modified island removal; fixed a basal freeze-on bug
Three significant changes in this commit:

(1) I added an option for prescribing a maximum surface slope in the Glissade driving stress calculation.
    This change removes the huge velocities seen with some Greenland input files at initialization.
(2) I modified subroutine remove_floating_islands such that any region not connected to grounded ice
    via dynamically active ice (as opposed to any ice of nonzero thickness) counts as an island.
(3) I fixed a bug in the calculation of new tracer values for the case of basal freeze-on.

Details follow:

(1) Maximum surface slope

In testing one of the Greenland 4-km input meshes (greenland_4km.glissade.10kyr.beta6.SSacab_c140919.nc),
I found huge initial ice speeds at a couple of coastal vertices, with a max of nearly 800 km/yr
at global location (260,253). The problematic vertices have a surface slope of ~0.4, with the
surface elevation decreasing by 1500 m over a distance of 4 km as the ice thins from ~1500 m to 0.
The velocity solver seems to be working correctly; huge ice speeds are expected with such a steep slope.

One possible solution would be to smooth the ice thickness and surface elevation in the input file.
Another would be to run the model for a short time (~1 yr) with a small timestep,
until the ice adjusts and the initial steep gradients are removed.
However, I don't think either solution is very robust.  There is still a chance of
wandering into a configuration with steep gradients in the future.

Instead, I adopted the fairly simple solution of allowing the user to impose a maximum surface slope.
This can be set in the config file by choosing max_slope in the parameters section.
The default value is 1.0, which basically amounts to not having a max_slope condition.
(I have not seen an input data file with a slope as large as 1.0.)

Here is how the maximum-slope condition is implemented:
Suppose max_slope = 0.10.  Consider 2 adjacent cells, 4 km apart, with a surface elevation
difference of 1 km and thus a slope ds/dx = -0.25 on the edge between the 2 cells.
In subroutine glissade_centered_gradient, ds/dx is reduced from -0.25 to -0.1 at this edge.
(Note that max_slope is a magnitude that is applied to gradient components of either sign.)
The limited value of -0.1 is used in the interpolation of the gradient from edges to vertices.
The vertex value can be up to sqrt(2)*max_slope, if there is a steep gradient in both
the x and y directions. The surface slope at the vertex is then inserted into the
finite-element sums in the Glissade dycore. So the largest slope that appears in the
equations of motion is not exactly equal to max_slope, but is of similar magnitude.

The implementation is similar in subroutines glissade_upstream_gradient and glissade_gradient_at_edges.
Note, however, that the upstream method implements the gradient as follows:
      df_dx(i,j) = 1.5d0 * df_dx_south - 0.5d0 * df_dx_south2
Thus if df_dx_south = max_slope = 0.1 and df_dx_north = -max_slope = -0.1,
then df_dx(i,j) = 2*max_slope = 0.2.
This means that when the topography is very rough, glissade_upstream_gradient may permit
velocities up to 2*max_slope.

I tested this change on a Greenland 4-km test problem with three difference input files.
The most problematic file, as mentioned above, was greenland_4km.glissade.10kyr.beta6.SSacab_c140919.nc,
which gives an initial speed of nearly 800 km/yr at global location (260, 253).
With max_slope = 0.1, the initial surface speed at this node is 30 km/yr, which is within the CFL limit
when dt = 0.05 year. Subsequent speeds at this and similar vertices are < 10 km/yr.
The maximum initial speed on the mesh is ~50 km/yr, which is asssociated with small beta
and fast sliding (a separate issue) rather than huge surface elevation gradients.

(2) I fixed an issue with subroutine remove_floating_islands.

This subroutine is used to remove floating ice that is disconnected from any grounded ice.
Previously, for purposes of this subroutine, ice of any nonzero thickness was counted
as active, providing a potential path to grounded ice. This allowed a single floating cell
to exist in dynamic isolation, causing problems.  Now, there must be a connection to grounded ice
through dynamically active cells (i.e., cells with thck > thklim), else free-floating cells are calved.

(3) I fixed a bug in the tracer calculation for the case of basal freeze-on.

There was an indexing error in the tracer_lsrf array, which determines the tracer value
in ice that is added at the bed. This error resulted in bad tracer values and/or seg faults.
This code had not been exercised before because I had been running with basal_mass_balance = 0
(no basal melt or freeze-on in continuity equation).

Some minor changes:

I fixed an issue with the max-speed diagnostic in the glide_diagnostics module.
Previously, if the cell-center thickness thck(i,j) = 0, then the velocity at vertex(i,j)
was excluded from consideration.  I changed thck to stagthck, so that any vertex
with stagthck(i,j) > 0 is now eligible for the max-speed designation.
On the Greenland coast, it is not unusual to have a vertex with stagthck(i,j) >> 0
adjacent to a cell with thck(i,j) = 0, with a steep gradient and large velocity at the vertex.
Such vertices are now included in the diagnostic.

In the commonly seen error diagnostic 'Remap, departure point out of bounds',
I added a few lines of code to print the global cell location, not just the local location.
This makes it easier to set idiag and jdiag for debugging.

Finally, I modified and cleaned up some calving diagnostics.

I tested these code revisions with LIVV on a Mac.  There are no answer changes for standard tests.