Made the diagnostic output interval more robust
Typically, the user specifies the interval of diagnostic output in the config file.
For example, suppose the timestep dt = 0.1 yr, and the diagnostic interval dt_diag = 0.5 yr.
Then diagnostic output will be written once every 5 timesteps, or 0.5 yr.

However, this was not implemented in a robust way. As a result, diagnostics could fail
to be written out due to rounding errors.  In recent CESM TG tests, for example,
I had dt = dt_diag = 0.5 years, but diagnostic output was written only for the first 46 years
of the simulation. After that, the model got confused because the ratio of model time to dt_diag
was not quite close enough to an integer.

With this commit, dt_diag is used at initialization to find ndiag = nint(dt_diag/dt).
Diagnostic output is then written every ndiag steps. This procedure is robust.

Note that if dt does not divide evenly into dt_diag, we simply round to the nearest integer
to get ndiag. There is no attempt to write diagnostics at uneven intervals.

Also note that the user can specify ndiag directly in the config file.
But if dt_diag > 0 is specified, it will override any previously specified ndiag.

The default is still to write no diagnostic output.