Created a time-varying beta_internal field separate from the fixed external beta
The beta field is prone to confusion when read from an external file.
On the one hand, we have the fixed external beta, and on the other hand, we have
the beta field used in the velocity solver to assemble the basal traction terms
in the matrix.  The latter field is weighted at each vertex by f_ground,
so beta = 0 for floating ice.

Previously, this was handled in Glissade by saving a 'beta_external' field
that was referenced by the velocity solver at the start of each solve.
Unfortunately, the weighted internal beta field was still being written
to restart files.  When it was read in after restart, exact restart was
broken, and there was a potential for very large velocities when ice advanced
into land-based regions with beta = 0.

With this commit, the weighted, time-varying beta field is called beta_internal,
and the fixed field is called beta.  The beta field is loadable from restart files.
The beta_internal field is optionally written to history files, but is not read
from restart files.

I considered using 'beta_external' to denote the fixed, external beta field,
but this would have led to complexities in recognizing a field called 'beta'
in the input file and interpreting it correctly.

Rather than changing beta to beta_internal all over Glam, I simply passed
model%velocity%beta to beta_external, and passed model%velocity%beta_internal to beta.

In addition, I implemented a minimum value beta = beta_grounded_min = 10 Pa yr/m
for land-based ice (f_ground > 0). I found that one of the Greenland input files
has beta = 0 for many land-based cells that are initially ice-free. Without a minimum
threshold for beta, this would cause problems if ice ever advanced to these cells.
For now, beta_grounded_min is declared in subroutine calcbeta.
At some point we could add it to the velocity derived type.

The standard LIVV tests were successful (on a Mac) with these changes.
I also verified that Glam answers are unchanged for a dome problem.