Added a higher-order basal sliding option based on melting/freezing at the bed
Glide has long had a basal sliding option with a no-slip BC where the bed is frozen,
and sliding allowed where the bed is at the pressure-melting-point temperature.
The higher-order dycores have many other basal sliding options but not this one.
Thanks to Jeremy Fyke for suggesting that we add this option for Glissade.

I added a new option, which_ho_babc = HO_BABC_BETA_TPMP = 1, where beta depends
on whether or not the bed temperature is at the pressure melting point.
To make room in the '1' slot, I moved the old HO_BABC_SIMPLE, which is serial-only
and not often used, to option 12.

The new option is implemented near the beginning of the Glissade velocity solve.
The model computes the basal pmp temperature for each ice-covered cell, interpolates
to vertices, and compares to the bed temperature (also interpolated to vertices).
Where these temperatures are sufficiently close, an integer field pmp_mask = 1;
elsewhere pmp_mask = 0.

The mask is passed to subroutine calcbeta and used to set beta at each vertex:
- Where T < Tpmp, beta has a large value of 1.e10 Pa yr/m, ensuring very low slip.
- Where T = Tpmp, beta = ho_beta_const, which is 10 Pa yr/m by default but can be
set by the user in the config file.

I ran a modified dome problem to make sure the new option is working.
It remains to be tested for a more realistic Greenland problem.

There are no answer changes for standard test problems.