Added a user-configurable minimum beta for grounded ice
In a previous commit (eb18c32), I modified subroutine calcbeta to ensure
that beta cannot be zero beneath grounded ice, but rather has a minimum value
of beta_grounded_min = 10 Pa yr/m. This fixed a problem with fast velocities
when reading a Greenland input file with beta = 0 in some regions.
However, it resulted in the confined-shelf test no longer being BFB.
(In the log I reported that results were BFB, but I must have overlooked
the confined-shelf failure.)

With this commit, beta_grounded_min becomes a user-configurable variable in
the glide_velocity derived type. It is passed into subroutine calcbeta from
the glissade solver as an optional argument. Its default value is 0.0, which means
that the confined-shelf results again are BFB. If using an input file with
zero values of beta in some regions, the user can switch to a safer value
(e.g., 10 Pa yr/m) by setting beta_grounded_min in the [parameters] section
of the config file. This is probably a good idea for CESM coupled runs.

After putting beta_grounded_min in the glide_velocity type, I put ho_beta_const
there too, for consistency. This resulted in a few minor Glam changes.
However, the Glam dycore does not use beta_grounded_min, because to do so
it would also need to compute f_ground, which would be a fair amount of work.

This commit fixes the BFB issue with the confined-shelf case.
Other standard LIVV cases are BFB.