Remove nan_mod
It is hard to handle NaNs in a way that works across all
compilers. Adding NAG support made this even harder. Rather than trying
to get it right, I am simply removing this. It was only being used in
dead code. In particular:

(1) libglide/glide_mask.F90 : glide_marine_margin_normal: This is not
called from anywhere in the code. I replaced NaN with 0.

(2) Three places in libglint/glint_interp.F90 : glint is no longer used
in CESM / ACME. Previously, all tasks initialized the 'global' output to
NaN, then the master initialized it to 0 (and it remained NaN on other
tasks). I have changed this so that it is initialized to 0 on all
tasks. This should be fine, since other tasks shouldn't be accessing the
value of 'global' (as documented, it is only valid on the master task).

In addition, I have removed the CISM_GNU option from the build: This was
in place in order to add a necessary compilation flag when compiling
nan_mod.F90; now that nan_mod is gone, this option can be removed, too.