Add target attribute
glimmer_sparse_slap: slap_default_options declares the 'base' dummy
argument as a target (and this is needed for the pointer assignment on
the last line). However, for this to be acceptable, the actual argument
must also have the 'target' attribute. Components of a derived type
cannot have the target attribute, so to accomplish this, I am giving the
actual object (the 'opt' object, in the caller) the 'target' attribute.

(An alternative would have been to make opt%base a pointer, but that
seems like more work.)