Cleaned up the glissade basal sliding options
I made a number of changes related to the basal sliding boundary condition:

* I revised option which_ho_babc = 1, which gives low beta and fast sliding
where the bed is at the pressure melting point temperature, with high beta
and slow sliding elsewhere. The low and high beta values are now given by
two user-configurable parameters, ho_beta_small and ho_beta_large, that are
part of the basal_physics derived type. The default value of ho_beta_small
is 1000 Pa yr/m.

* A 2D mask, bpmp_mask, is now set to 1 at vertices where the interpolated
bed temperature is greater than the interpolated pmp temperature minus a small
threshold.  The threshold is set to 0.001 degrees by default.
At some point we might want to consider a linear ramp in beta as Tbed approaches
the pmp temperature.

* To avoid a proliferation of input arguments in subroutine calcbeta,
I moved several parameters and fields to the basal_physics derived type:
ho_beta_large, ho_beta_small, and bpmp_mask as mentioned above, and also
ho_beta_constant, beta_grounded_min, p_ocean_penetration and mintauf.
All these parameters and fields are now passed in as part of 'basal_physics'
instead of having to be passed separately. This leads to some simplification
in both the glam and glissade velocity solvers.

* I added a line of code to copy the beta_internal field to beta whenever
beta is not read from an external file.  Recall that beta_internal is the
field actually used for the velocity solve and is weighted by f_ground.
If beta is read externally, it needs to remain fixed, but otherwise there
is no harm in making it equal to beta_internal.  Thus, users who add 'beta'
to the list of output fields will get what they expect, without having to
remember to list 'beta_internal' instead.

* I moved the halo update for beta_internal from glissade_velo_higher_solve
to calcbeta.

* I made pmp_offset a user-configurable parameter.  This determines how far
below the pressure melting point the bed temperature is set at initialization
for the temp_init = 2 option (a linear profile between surface and bed).
The default remains pmp_offset = 2 degrees.

* I made bpmp an optional output field in glide_vars.def. This makes it easier
to visualize where the bed is at pressure melting point.

* I removed an unused basal sliding option, which_ho_babc = HO_BABC_BETA_BWAT = 3.
Another option may be added in the '3' slot later.

* I slightly changed the names of some of the other which_ho_babc options.

* I removed the unused basalproc and glide_grnd derived types.
(Nothing was left in basalproc after mintauf was moved to basal_physics.)

* An unrelated change: In glide_vars.def, I changed units of enthalpy from
'J/m^3' to "J/m3' so that the units can be parsed without an error in the
latex documentation.

I verified using LIVV on a Mac that answers are BFB for standard tests.
Answers are also BFB for a dome test problem using glam.
In the case of the stream test, keeping the answers BFB requires scaling
mintauf (as before) inside glissade_velo_higher_solver. I would prefer to
remove that scaling, and will do so in a future non-BFB commit.