Changed the logic for applying uvel_extend and vvel_extend at initialization
For problems with global periodic BC, CISM uses the fields uvel_extend and vvel_extend
to ensure exact restart.  These fields are like uvel and vvel, but with an extra row
along the north and east edges to ensure that all velocity values are included.

Previously, uvel_extend and vvel_extend were read and copied to uvel and vvel only at restart.
In some idealized tests, however, these fields need to be copied to uvel and vvel at the initial
time (i.e., before any restarts) to enforce a Dirichlet BC on velocity at a global boundary.
This is the case for some calving tests carried out recently by Ryan Whitcomb (U. Michigan).

With this commit, the logic has changed so that uvel_extend and vvel_extend, if present
in the input file (either a standard input file or a restart file), are always copied
to uvel and vvel at time t = tstart.

Thanks to Ryan for discovering and helping to diagnose the problem.