Fixed a bug in the local calculation of bfricflx
Recently, a new option (which_ho_assemble_bfric = HO_ASSEMBLE_BFRIC_LOCAL = 1) was added
for the computation of bfricflx, the frictional heating term associated with basal sliding.
It is natural to compute bfricflx at velocity points (i.e., vertices), but the heating should
be applied at cell centers.  The previous implementation of this option did not include
an interpolation from vertices to cell centers.

With this commit, bfricflx is interpolated to cells. The results look correct for a dome problem.

I verified (with LIVV on a Mac) that the answers are BFB for standard tests.