Set Tbed = pmptemp in the presence of basal water
I changed the basal temperature calculation so that Tbed is fixed at pmptemp
(the pressure melting point temperature) if either
(1) Tbed is already very close to pmptemp, or
(2) Basal water is present (bwat > 0).

Previously, Tbed was held fixed at pmptemp only in case (1).
Thanks to Marion Bougamont for pointing out that (2) should also apply.

"Very close to pmptemp" was previously defined as within 0.001 deg C of pmptemp,
with 0.001 being a hardwired constant.
Now it is defined in terms of a user-configurable parameter, pmp_threshold.
If Tbed >= pmptemp - pmp_threshold, then the bed is treated as thawed.
The default value of pmp_threshold is 0.001 deg C, so model behavior should not change
unless the user changes pmp_threshold in the config file.

Another minor change: When Tbed = pmptemp but the ice wants to cool (i.e., basal freeze-on would occur
if basal water were present, but bwat = 0), Tbed is set to pmptemp - 2*pmp_threshold.
Previously, it was set to pmptemp - 0.01 deg. This change will slightly change the answers
for problems with a bed that alternates between thawed and frozen, but this is not the
case for the standard LIVV tests.

I ran some dome tests and verified that the new code works as expected when basal water is present.
NOTE: Users may want to experiment with values of pmp_threshold greater than 0.001 deg C.
With the dome problem, I found that even if Tbed is set to pmptemp in each grid cell during the
thermal calculation, we can have Tbed < pmptemp by ~0.01 deg C at some vertices after interpolation.

I verified with LIVV on a Mac that standard test results are BFB.