Allowed the thermal solve to be done before and/or after transport; modified glissade_tstep
The vertical thermal solve in CISM has always been done before the transport solve.
This may not be ideal. For example, suppose we are using a sliding law for which
the basal traction depends on whether the bed is thawed or frozen.
If T_bed is computed before transport (during the thermal solve) but T_pmp is computed
after transport (during the diagnostic solve), then we cannot accurately diagnose
whether the bed is frozen or thawed. Using the PISM pseudoplastic BC with a
basal-temperature-dependent effective pressure, this can lead to oscillations
in basal velocity in some grid cells; T_bed can flip back and forth between values
above and below T_pmp.

One good solution is to move the thermal solve after the transport.
Another is to do the thermal solve twice, before and after the transport,
using a time step of dt/2 each time.  The second option may be more accurate,
since the transport scheme will then operate on temperatures that exist
halfway through the time step.

With this commit, a new config option called which_ho_thermal_timestep controls
when the thermal solve is done relative to the transport solve. The options are:

      integer, parameter :: HO_THERMAL_BEFORE_TRANSPORT = 0
      integer, parameter :: HO_THERMAL_AFTER_TRANSPORT = 1
      integer, parameter :: HO_THERMAL_SPLIT_TIMESTEP = 2

The default is 0 (ensuring no change in results with default parameters).
Options 1 and 2 greatly reduce oscillations in basal velocity in coastal Greenland
when using a pseudoplastic basal BC with a basal temperature threshold.
For dt = 0.5 yr, the largest remaining oscillations, around vertex (92,363), are
associated not with variations in beta but with steep surface elevation gradients.
Setting dt = 0.25 yr reduces these oscillations.

To implement this option with minimal code duplication, I created a new subroutine,
glissade_thermal_solve, that can be called from two different places in glissade_tstep,
with a timestep of either dttem or dttem/2.

For consistency, I also created driver subroutines called glissade_transport_solve,
glissade_calving_solve and glissade_isostasy_solve, all in the glissade module.
This makes glissade_tstep shorter and easier to follow.

Also, I removed the global basal pmp calculation from glissade_velo_higher_solve
and put it near the beginning of glissade_therm_driver. This ensures that T_pmp and T_bed
are consistent with each other, even if they are not up to date with the ice thickness.
On the first diagnostic solve, the basal pmp is calculated before calling the
velocity solver (since the thermal solve has not been called yet).

Answers are BFB for standard tests with default parameters (verified with LIVV on Mac).