Removed scaling parameter of -1 for the basal heat flux
Previously, CISM used a scaling parameter scale_bflx = -1.
This parameter was used to flip the sign from the input geothermal heat flux
(assumed to be defined as positive upward) to the CISM sign convention
(positive downward).

Most CISM input files, however, use the positive-down convention, in which
case no sign flip is needed.  For this reason I removed the scaling parameter,
which means that henceforth, we will assume that the input geothermal heat
flux is positive down (so the values will be negative).

NOTE: Any input files with a geothermal flux (bheatflx) using the positive-up
convention will need to have their sign reversed, for either Glide or Glissade.

I added an initialization check that aborts the code with an error message
if the max value of the input geothermal flux is positive on any processor.

I verified that results are BFB for standard LIVV tests on a Mac.