Added a temperature init option using advective-diffusive balance
Until now, CISM has used two main temperature initialization options:

(1) TEMP_INIT_ARTM:   Initialize each column to the surface air temperature
(2) TEMP_INIT_LINEAR: Initialize each column to a linear profile, from the
    surface air temperature at the top to a value close to Tpmp at the bed

For (1), it takes a long time to spin up temperatures near the bed for thick ice.
We have been using (2) for Greenland spinups, but this method has its own issues.
It gives temperatures that are too warm near the surface in the ice sheet interior,
where downward cold advection competes with upward diffusive heat fluxes to
set the ice temperature. If we start with temperatures that are too warm higher up
in the column, then the transfer of heat away from the bed is slowed, and the
bed can thaw in regions where it should be frozen.

Near the divide, horizontal advection is small and the temperature is
determined by a vertical balance between advection and diffusion:

           w dT/dz = (k/rhoi*ci) * d2T/dz2

where w = vertical velocity = -SMB.

Given T = Tair at the top surface and dT/dz = -F_geo/k at the bed, this equation
has an exact solution (see Cuffey and Paterson, 2010, Section 9.5.1).
The exact solution assumes w < 0 and hence applies only where SMB > 0.
This solution is a reasonable initial guess for temperature in the ice interior.

With this commmit there is a new option, temp_init = TEMP_INIT_ADVECTIVE_DIFFUSIVE = 3.
The temperature profile is first initialized as in method (2).
Then, where SMB > 0, we set the initial temperature in each layer to the linear value
or the advective-diffusive value, whichever is lower.

In this way we obtain more realistic temperature profiles near the divide,
allowing the temperature to spin up more quickly than method (1), but without
as much bed thawing as in method (2).

For debugging purposes, I moved the call to glide_init_diag before the call to
glissade_init_therm, so that the indices of the chosen diagnostic point are
available in glissade_init_therm.

I also backed off on a previous change in glissade_therm, whereby thin ice (thck < thklim_temp)
is given a temperature profile based on the temp_init option.  Instead, thin ice is
initialized to min(0.0, artm) throughout the column.  It would be difficult to apply
the new temp_init option to thin ice during runtime with a time-varying acab.

I verified that initial temperature profiles look as expected for Greenland simulations
with a prescribed SMB (acab).  If an acab field is present in the input file, it can be
used for this temp_init option, even if the actual SMB seen by CISM is computed at runtime.

Since this option is not the default, results of standard tests are BFB
(verified by LIVV on a Mac).