Changed the criteria for which_ho_gradient_margin = 1
The purpose of which_ho_gradient_margin = 1 is to provide a gradient computation
that works well at either land-terminating or marine-terminating margins.

Before this commit, gradients were computed at edges where either
(1) Both adjacent cells are ice-covered.
(2) One cell is ice-covered land, and lies higher in elevation than the adjacent cell.
    The elevation condition was included so that nunataks would not drive large gradients.
At ocean margins, where an ice-covered non-land cell is adjacent to an ice-free ocean cell,
the gradient was masked out. Thus the gradient at the edge of ice shelves is zero,
to represent a vertical ice cliff.

This commit modifies criterion (2), such that the gradient is computed if one ice-covered
cell is grounded, whether or not it is on land.  So the gradient is now computed
for grounded marine-based cells adjacent to ice-free ocean cells.
The new logic uses the glissade floating_mask instead of land_mask.

This addresses a problem case that can turn up in Greenland simulations:
If a grounded marine-based cell is entirely surrounded by ice-free ocean,
we want ice to flow to the ocean.  Otherwise, if the SMB > 0, ice can accumulate
in the grounded cell indefinitely and never move.

I verified that this change fixes the problem of an indefinitely growing cell
with global indices i = 185, j = 155 on the 4 km Greenland mesh.

While changing the edge-masking subroutine, I removed some older deprecated code.
Also, I changed the name of HO_GRADIENT_MARGIN_ICE_LAND to HO_GRADIENT_MARGIN_HYBRID.
"Hybrid" better captures the idea that we have one kind of BC for land-terminating margins
and another kind of BC for marine-terminating margins. The option number is still 1.

NOTE: This change modifies results for the confined-shelf test.  Previously, the confined-shelf
results were the same for which_ho_gradient_margin option 1 as for option 2.  New results are BFB
for option 2, but have changed for option 1 (because there are some grounded marine-based cells
adjacent to ice-free ocean cells).

For this reason, I am checking in a new version of shelf_confined.config, in which
which_ho_gradient_margin = 2 instead of 1. With this change, I verified that the
confined-shelf results are BFB compared to the benchmark.

If you are running LIVV with an older benchmark and the confined-shelf test fails,
you may have to change the config file manually in ../cism_reg/tests/higher-order/shelf.
Simply add which_ho_gradient_margin = 2 in section [ho_options].

Other standard tests are BFB without any changes (confirmed by LIVV on a Mac).