Reverted from bmlt_ground and bmlt_float to one bmlt field
In an earlier commit, the basal melting field (bmlt) was split into
two separate fields, bmlt_ground and bmlt_float.  This was done partly
to facilitate separate diagnosis of the two forms of basal melting,
which generally have different physical causes.

However, the separation breaks interfaces with external dycores and is
not really necessary, since it is easy to look at bmlt for floating ice
only (or grounded ice only) using floating_mask.

With this commit I am reverting to a single field, bmlt.
Within glissade_therm.F90, there are local fields bmlt_ground and bmlt_float
that are computed by two separate subroutines, but these local fields
are summed to give the total bmlt for output.

Answers are BFB for standard tests (verified by LIVV on a Mac) and for
standalone runs generally.

[NOTE TO BILL SACKS:]
However, the output fluxes computed by Glad could differ, because Glad was
previously accumulating basal melting from grounded ice only, and now it
is accumulating basal melting from both grounded and floating ice.
Even if the ocean is not supplying a flux for basal melting, there is a
possibility of internal melting in floating ice.
This internal melting would now be (correctly) routed to the ocean,
whereas previously it would have been ignored.