Removed obsolete glissade_temp and glissade_enthalpy modules
Some time ago, the glissade_temp and glissade_enthalpy modules were
combined into a single module, glissade_therm.  At that time I verified
that the new glissade_therm module gave the same answers as the old
glissade_temp module for the standard vertical temperature solve.

I did not test as thoroughly whether the new glissade_therm module
was working as intended for a vertical enthalpy solve.
This was going to be done by the CU developers of the enthalpy scheme,
but slipped through the cracks.  As far as I know, the enthalpy scheme in
glissade_therm is working, but it needs more testing and is not yet
scientifically supported.

In order to avoid confusion for the 2.1 release, I am removing the
old glissade_temp and glissade_enthalpy modules. At some point,
someone can fully test the enthalpy scheme as it exists in glissade_therm.