Remove surftemp from glide_vars.def and regenerate default i/o files
There are problems if you try to load or output surftemp from a glide
run. Bill Lipscomb and I can't see a straightforward way to prevent
the loading or outputting of surftemp just for glide, so for now we're
simply removing this as a possible input / output. artm can be used
instead.