Modified the power-law basal BC for consistency with MISMIP+
One of the basal-sliding options for MISMIP+ is a simple Weertman-type power law
of the form tau_b = C * u^(1/m), where C and m are prescribed constants.
Until now, CISM hasn't supported this option, but had a slightly more
complicated option (which_ho_babc = HO_BABC_POWERLAW = 9) that includes
dependence on effective pressure.

With this commit, option 9 becomes the simpler power law specified for MISMIP+.
This will allow comparison of MISMIP+ results between CISM and MPAS LI at similar resolutions.
(I'm running these comparisons now.)

I'm not sure that the old option 9 will ever be used, but for now I switched it
to option 13.

A couple of minor changes: Swapped indices (1 <-> 2) for the bmlt_float options,
and changed a threshold in the f_ground calculation.

Standard tests are BFB (confirmed by LIVV on a Mac)

These changes are on Gunter's misomip branch and are not part of the 2.1 release.