Changes for MISMIP+ Ice2 forcing
MISMIP+ experiments Ice2r, Ice2ra and Ice2rr specify a large basal melt rate
beneath floating ice with x > 480 m.  The melt rate is computed in glissade_therm
and applied in glissade_transport. Previously, a nonzero melt rate was computed only
in cells with ice present at the time glissade_therm is called. Now, it is
computed for ice-free ocean cells too, and then applied if these cells acquire ice
(via advection) in glissade_transport.  This makes the results much cleaner.
Similarly, SMB is now set to zero for ice-free ocean cells. A parameter called
bmlt_float_xlim is used to turn off melt for x < 480 km.

With these changes, CISM results for Ice2r are very close to MPAS LI results.
I made a similar set of changes in MPAS LI (to be checked in shortly).

Other changes:
- Moved f_ground calculation to glissade_diagnostic_variable_solve, so that f_ground
  is always available as a diagnostic
- Simplified which_ho_babc = 11 by combining three config parameters into one

Results of standard tests are BFB (confirmed by LIVV on a Mac).
The changes are on Gunter's misomip branch and are not part of the 2.1 release.