Fixed a minor bug in incremental remapping
There is logic in the IR code to test whether the slope of a departure trajectory
is nonzero.  In one line I changed 'puny' to 0.d0. This is needed to give a
correct answer when the slope is less than puny (a very small number) but > 0, as
can happen when prescribed trajectories are almost exactly along the x-axis.

Results are BFB for standard tests. In rare cases the results could change,
and would now be correct instead of incorrect.