Minor change in remapping algorithm
This is a one-line change of a threshold from 'puny' to 'zero' in the geometry portion
of the incremental remapping algorithm.  In cases where the flow is very nearly along
a coordinate axis, the appropriate threshold is zero.

The change is BFB for standard tests (verified by LIVV on a Mac).