Changed 'scale_factor' to 'area_factor' in stereographic derived type
For Greenland initMIP diagnostics, I recently added an area scale factor
that describes the area distortion on a stereographic grid. It is part
of a derived type for stereographic projections.

With this commit, I am simply changing the variable name from 'scale_factor'
to 'area_factor' to avoid confusion with the scale factor in netCDF files.

Answers are BFB.