Added option to reduce effective pressure when bmlt > 0
There are several options for reducing effective pressure when the bed is thawed
or basal water is present. This is a new one, based on the idea that effective
pressure should decrease when there is melt at the bed.

I added a config option in glide_types and some new logic in the basal traction subroutine.
However, the new option remains to be tested and tuned.