Fixes for the stereographic projection area factor
The code was failing in the case that the input file contains projection mapping info.
The reason was that the subroutine that computes area_factor was called just after
reading glide_config, before reading the input file.  As a result, area_factor
was not being allocated and computed.

The fix is to compute the area factor after reading the input file.
This is now done in both glissade_initialise and glide_initialise.
(Previously, there was no call in glide_initialise.)
The area_factor array is now allocated in the same subroutine where it is computed,
instead of earlier in glide_allocarr.

Also, I cleaned up the area_factor calculation a bit.
I added code to compute area_factor correctly in the case lon_c = 90 degrees.
Since this code has not yet been tested, it triggers a non-fatal warning message.

Changes are BFB.