Add an 'internal_time' output variable
There are now two time variables on output files:

- 'internal_time' stores CISM's internal time variable, relative to the
  starting point of this simulation plus the amount of time that elapsed
  leading up to the restart file (if any)

- 'time' stores a more externally-useful version of time; for climate
  model runs, this should agree with the climate model's time

The reason for having these two different variables is: There can be
roundoff-level differences upon restart if the internal time variable is
modified upon restart. So it is important that this variable is
maintained exactly as is when reading a restart file. However,
maintaining this value as is isn't always the right thing to do in terms
of the climate model's time, so we sometimes need a separate variable to
represent that time.

It's possible that, with some more thought, these two variables could be
combined into one. But having two variables seemed like the easiest and
least error-prone solution for now.

This commit is just the first step: The next step will be allowing
'time' to differ from 'internal_time' (currently they are identical).