Fixed isostasy restart, introducted tstep_count variable
In the previous commit, a mod statement with real arguments was used to determine
when it was time to update the lithospheric load for the isostasy.  This method
turned out not to be robust, because of small rounding errors in the time variable.

With this commit, the model determines the update time based on an integer,
model%isostasy%nlith = nint(model%isostasy%period / model%numerics%tinc).
The load is updated on time step 1, and again every nlith steps thereafter.
For instance, if dt = 1 yr and isostasy%period = 500 yr, then updates will be done
on steps 1, 501, 1001, etc.  This is true whether the model runs continuously,
or stops and restarts.

To make this logic work, I introduced a new member of the numerics derived type:
tstep_count = the number of time steps since the simulation started.
This counter is set to zero at initialization and then is incremented in cism_run_dycore
before each call to the dycore (glide, glissade or external).
For a restart, tstep_count is initialized as nint(model%numerics%tstart / model%numerics%tinc).

Surprisingly, such a counter did not exist before. There was a counter called
model%numerics%timecounter, but this counter was set to 0 on restart.
I removed model%numerics%timecounter to reduce confusion.
Calls to glide_write_diagnostics now pass in model%numerics%tstep_count instead.

Also, I changed the name of the lithosphere update period in the config file from
'period' to 'lithosphere_period', to reduce the likelihood of name conflicts.

I verified that isostasy restarts exactly for standalone runs on my Mac,
with update periods both longer and shorter than the restart interval.
Bill S. can now do the same for CESM/CISM tests.

Changes are BFB except for restarted runs with isostasy.